Minor bug fixes and updates.
Capitalized the 'P' in "PlayerListener."
Fixed bug where using too high of a value gave infinite glow.
Providing a value of -1 provides infinite glow.
Proving a value of 0 informs the user of the clearglow command.
Providing a negative value other than negative informs the player of the
proper values.
Providing a value > 1500 informs the player about the infinite glow
value of -1.