New config command functionality and a modified config file!
Created a DataVerification abstract class that all data verification
classes inherit from.
Updated Color and Duration Verifications to inherit from the new
abstract class.
Created an EventCauseVerification class and a SettingValueVerification
class.
Created a new default config file, which utilizes enums and other data
for an easier time adding future features.
Updated the listener to use the new config format.
Added a removeglowconfig command to remove configurations.
Removed the setGlowConfig and replaced it with the configGlow command.
The new glowConfig command, that modifies individual glow attributes.
Updated the plugin to version 0.9.0.
WARNING: This version of the plugin is not tested!